fetching traces
